修改项目列表bug(修改条件查询老师和企业老师),和修改根据id 查询bug(项目不存在时候抛出异常) 和 修改Controller请求参数方式

+++ cms-models/cms-models-xiangmu/src/main/java/com/xk/domain/vo/detail/Student.java
@@ -19,7 +19,7 @@ public class Student {
     /**
      * 是否负责人 , 0为不是, 1是(只能存在一个)
      */
-    private Long isPrincipal;
+    private Integer isPrincipal;
     /**
      * 分工
      */
